[gn] Explicitly include common.gni in build files that use vars from it (#2973)
runtime/BUILD.gn and ..core/BUILD.gn use flutter_runtime_mode which
is defined in common.gni. If the build comes through common/BUILD.gn
this is imported (implicity) but builds that depend on specific
files, like the Fuchsia build, won't have this imported.  This
imports the gni from the build files that use it instead of relying
on it already being there.