Search multiple paths when loading deferred component .so files. (#23849)
This is a partial revert of 7c19824c6d08bb8e3fc4217058178d42d451e26b

On some devices we still need to the original search paths approach
because dlopen with just the base file name doesn't work. We're
combining both approaches now, adding the base filename as the first
entry in the searchPaths.